Enhance installation scripts for VS Code and Docker with user prompts and final messages; remove obsolete finalise script

+++ .setup/scripts/001-vscode.sh
@@ -3,18 +3,26 @@
 # Error handling
 set -e
 
-# Install dependencies for VS Code repo
-sudo apt install -y \
-  wget \
-  gpg
+# Prompt for install
+echo
+echo -ne "${YELLOW}Install VSCode? [Y]: ${NC}"
+read -r install_vscode < /dev/tty
+echo
+install_vscode=${install_vscode:-Y}
+if [[ $install_vscode =~ ^[Yy]$ ]]; then
 
-# Import Microsoft GPG key
-wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
-sudo install -D -o root -g root -m 644 microsoft.gpg /usr/share/keyrings/microsoft.gpg
-rm -f microsoft.gpg
+  # Install dependencies for VS Code repo
+  sudo apt install -y \
+    wget \
+    gpg
 
-# Add reference to upstream repository
-sudo tee /etc/apt/sources.list.d/vscode.sources > /dev/null <<_EOF_
+  # Import Microsoft GPG key
+  wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
+  sudo install -D -o root -g root -m 644 microsoft.gpg /usr/share/keyrings/microsoft.gpg
+  rm -f microsoft.gpg
+
+  # Add reference to upstream repository
+  sudo tee /etc/apt/sources.list.d/vscode.sources > /dev/null <<_EOF_
 Types: deb
 URIs: https://packages.microsoft.com/repos/code
 Suites: stable
@@ -23,7 +31,9 @@ Architectures: amd64,arm64,armhf
 Signed-By: /usr/share/keyrings/microsoft.gpg
 _EOF_
 
-# Install VS Code
-sudo apt install -y apt-transport-https
-sudo apt update -y
-sudo apt install -y code
+  # Install VS Code
+  sudo apt install -y apt-transport-https
+  sudo apt update -y
+  sudo apt install -y code
+
+fi
